+++ recipes/config.rb
@@ -18,20 +18,20 @@
 #
 
 # always keep aws.conf updated
-template "#{node['aws-cwlogs']['path']}/etc/aws.conf" do
+template "#{node['aws_cwlogs']['path']}/etc/aws.conf" do
    source 'aws.conf.erb'
    owner 'root'
    group 'root'
    mode 0600
    variables ({
-      :awsRegion => node['aws-cwlogs']['region'],
-      :awsAccessKey => node['aws-cwlogs']['aws_access_key_id'],
-      :awsSecretKey => node['aws-cwlogs']['aws_secret_access_key']
+      :awsRegion => node['aws_cwlogs']['region'],
+      :awsAccessKey => node['aws_cwlogs']['aws_access_key_id'],
+      :awsSecretKey => node['aws_cwlogs']['aws_secret_access_key']
    })
 end
 
 # always keep logging.conf updated
-template "#{node['aws-cwlogs']['path']}/etc/logging.conf" do
+template "#{node['aws_cwlogs']['path']}/etc/logging.conf" do
    source 'logging.conf.erb'
    owner 'root'
    group 'root'
@@ -40,18 +40,17 @@ end
 
 # always generate awslogs.conf based on default
 # attributes related to log files
-template "#{node['aws-cwlogs']['path']}/etc/awslogs.conf" do
+template "#{node['aws_cwlogs']['path']}/etc/awslogs.conf" do
    source 'awslogs.conf.erb'
    owner 'root'
    group 'root'
    mode 0644
-   variables ({
-      :logFiles => node['aws-cwlogs']['log_files']
-   })
 end
 
 # always restart aws cloudwatch logs agent
 # after the configuration files were updated
 service 'awslogs' do
-   action [:restart]
+   action [:enable, :restart]
+   supports :restart => true, :status => true, :start => true, :stop => true
+   subscribes :restart, "template [#{node['aws_cwlogs']['path']}/etc/awslogs.conf]", :delayed
 end