+++ recipes/default.rb
@@ -17,22 +17,22 @@
 # limitations under the License.
 #
 
-if node['aws-cwlogs']['region'].nil?
+if node['aws_cwlogs']['region'].nil?
    log('AWS Region is necessary for this cookbook.') { level :error }
    return
 end
 
-if node['aws-cwlogs']['aws_access_key_id'].nil?
+if node['aws_cwlogs']['aws_access_key_id'].nil?
    log('AWS Access Key is necessary for this cookbook.') { level :error }
    return
 end
 
-if node['aws-cwlogs']['aws_secret_access_key'].nil?
+if node['aws_cwlogs']['aws_secret_access_key'].nil?
    log('AWS Secret Access Key is necessary for this cookbook.') { level :error }
    return
 end
 
 # only install if it isn't installed
-include_recipe 'aws-cloudwatchlogs::install' unless ::File.exist?(node['aws-cwlogs']['path'])
+include_recipe 'aws-cloudwatchlogs::install' unless ::File.exist?(node['aws_cwlogs']['path'])
 # always reconfigure aws cloudwatch logs configuration files
 include_recipe 'aws-cloudwatchlogs::config'