+++ recipes/install.rb
@@ -18,23 +18,23 @@
 #
 
 # create base directory of agent even if it isn't installed yet
-directory "#{node['aws-cwlogs']['path']}/etc" do
+directory "#{node['aws_cwlogs']['path']}/etc" do
   recursive true
 end
 
-template "#{node['aws-cwlogs']['path']}/etc/aws.conf" do
+template "#{node['aws_cwlogs']['path']}/etc/aws.conf" do
    source 'aws.conf.erb'
    owner 'root'
    group 'root'
    mode 0600
    variables ({
-      :awsRegion => node['aws-cwlogs']['region'],
-      :awsAccessKey => node['aws-cwlogs']['aws_access_key_id'],
-      :awsSecretKey => node['aws-cwlogs']['aws_secret_access_key']
+      :awsRegion => node['aws_cwlogs']['region'],
+      :awsAccessKey => node['aws_cwlogs']['aws_access_key_id'],
+      :awsSecretKey => node['aws_cwlogs']['aws_secret_access_key']
    })
 end
 
-template "#{node['aws-cwlogs']['path']}/etc/logging.conf" do
+template "#{node['aws_cwlogs']['path']}/etc/logging.conf" do
    source 'logging.conf.erb'
    owner 'root'
    group 'root'
@@ -46,13 +46,10 @@ template '/tmp/awslogs.cfg' do
    owner 'root'
    group 'root'
    mode 0644
-   variables ({
-      :logFiles => node['aws-cwlogs']['log_files']
-   })
 end
 
 # download setup script that will install aws cloudwatch logs agent
-remote_file "#{node['aws-cwlogs']['path']}/awslogs-agent-setup.py" do
+remote_file "#{node['aws_cwlogs']['path']}/awslogs-agent-setup.py" do
    source 'https://s3.amazonaws.com/aws-cloudwatch/downloads/latest/awslogs-agent-setup.py'
    owner 'root'
    group 'root'
@@ -61,7 +58,7 @@ end
 
 # install aws cloudwatch logs agent
 execute 'Install CloudWatch Logs Agent' do
-   command "#{node['aws-cwlogs']['path']}/awslogs-agent-setup.py -n -r #{node['aws-cwlogs']['region']} -c /tmp/awslogs.cfg"
+   command "#{node['aws_cwlogs']['path']}/awslogs-agent-setup.py -n -r #{node['aws_cwlogs']['region']} -c /tmp/awslogs.cfg"
    not_if { system 'pgrep -f awslogs' }
 end
 
@@ -69,4 +66,5 @@ end
 # the agent will run with the custom configurations
 service 'awslogs' do
    action [:enable, :restart]
+   supports :restart => true, :status => true, :start => true, :stop => true
 end