+++ CHANGELOG.md
@@ -1,8 +1,13 @@
 Changes
 =======
 
+# 1.1.5 / 2017-01-31
+* Implementing a better way to handle AWS CloudWatch Logs configurations,
+  now each log has a configuration file that is provisioned via Chef Resources.
+
 # 1.1.4 / 2016-11-07
-* Implementing `logging_config_file` configuration, that overrides the default logging configuration to a WARNING value.
+* Implementing `logging_config_file` configuration, that overrides the default
+  logging configuration to a WARNING value.
 
 # 1.1.3 / 2016-10-31
 * Fixing `aws_access_key_id` and `aws_secret_access_key` attributes values.